# openapi.yaml
openapi: 3.0.3
info:
  title: Transparency Dashboard API
  description: API for tracking grants and spending on-chain for organizational transparency
  version: 1.0.0
  contact:
    name: Ethereum Community Foundation
servers:
  - url: /api/v1
    description: API server
paths:
  /api/v1/treasury:
    get:
      summary: Get treasury assets
      description: Returns the treasury assets, wallets, and total values for the organization
      operationId: getTreasury
      tags:
        - Treasury
      responses:
        '200':
          description: Successfully retrieved treasury information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreasuryResponse'
              examples:
                success:
                  summary: Example treasury response
                  value:
                    organizationName: "Ethereum Community Foundation"
                    assets:
                      - name: "ETH"
                        amount: "1250.5"
                        usdWorth: 3126250.0
                        ethWorth: "1250.5"
                      - name: "USDC"
                        amount: "500000.0"
                        usdWorth: 500000.0
                        ethWorth: "200.0"
                    wallets:
                      - address: "0x1234567890abcdef1234567890abcdef12345678"
                        etherscanLink: "https://etherscan.io/address/0x1234567890abcdef1234567890abcdef12345678"
                    totalValueUsd: 3626250.0
                    totalValueEth: "1450.5"
                    lastUpdated: "2023-12-15T10:30:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    TreasuryResponse:
      type: object
      properties:
        organizationName:
          type: string
          description: Name of the organization
          example: "Ethereum Community Foundation"
        assets:
          type: array
          description: List of treasury assets
          items:
            $ref: '#/components/schemas/Asset'
        walelts:
          type: array
          description: List of organization wallets
          items:
            $ref: '#/components/schemas/Wallet'
        totalValueUsd:
          type: number
          description: Total USD value of all treasury assets
          example: 3626250.0
        totalValueEth:
          type: string
          description: Total ETH equivalent value (as string to preserve precision)
          example: "1450.5"
        lastUpdated:
          type: string
          format: date-time
          description: Timestamp when the treasury data was last updated
          example: "2023-12-15T10:30:00Z"
      required:
        - organizationName
        - assets
        - wallets
        - totalValueUsd
        - totalValueEth
        - lastUpdated
    Asset:
      type: object
      properties:
        name:
          type: string
          description: Name of the asset
          example: "ETH"
        amount:
          type: string
          description: Amount of the asset (as string to preserve precision)
          example: "1250.5"
        usdWorth:
          type: number
          description: USD value of the asset
          example: 3126250.0
        ethWorth:
          type: string
          description: ETH equivalent value (as string to preserve precision)
          example: "1250.5"
      required:
        - name
        - amount
        - usdWorth
        - ethWorth
    Wallet:
      type: object
      properties:
        address:
          type: string
          description: Ethereum wallet address
          example: "0x1234567890abcdef1234567890abcdef12345678"
        etherscanLink:
          type: string
          format: uri
          description: Link to view the wallet on Etherscan
          example: "https://etherscan.io/address/0x1234567890abcdef1234567890abcdef12345678"
      required:
        - address
        - etherscanLink
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to fetch treasury data"
        code:
          type: string
          description: Error code
          example: "TREASURY_FETCH_ERROR"
      required:
        - error
        - code
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...